Beginnings of menu support.
The XKeyCodes class is there because I intended to do this through
sending keyboard input, but there doesn't seem to be a reliable
way to do that in wayland, even with AT-SPI enabled.

This makes menu support rather complex, but it should work in
theory once everything it uses is implemented.