Fixes for wxwidget combo boxes.
The wxwidget combo box menu items do not have the selectable state,
even though they can be selected. They also always have the
showing state, meaning that it's not possible to tell whether a
combo box menu is opened. (Oddly, the menu child of the combo box
isn't created until the first time it's opened, so things seem
fine until then.)

Fortunately, wxwidgets also shows a separate window containing a
menu, which is consistent with gtk. So ignoring menu children of
combo boxes works around this.

Unfortunately, this still means that if there are multiple combo
boxes, we can't tell which one is open, which is required to
be able to dismiss it through AT-SPI. So we will need keyboard
event synthesis to handle this properly.