Add Tmds.Dbus as a submodule.
I haven't worked out how to do any build integration yet, but at
least it'll be possible to figure out what version was in a given
release.

Ideally, upstream would have a work-around for the bug affecting
Mono (which is probably a Mono bug but I haven't been able to
figure out what's going on exactly), and we wouldn't need this.