Simplify descent code and check for active ancestor.
This changes the behavior of is_active_modal so it only has an
effect if it's inside the "current" toplevel, and the deepest
is_active_modal element is used. This allows descent code to work
by just setting is_current_view to true or false.

It has the bonus of doing the right thing when the current view
is destroyed.