Uia: Remove assumption about MSAA parent id's.
The parent of a non-zero idChild is not necessarily the hwnd.

This is incomplete. We need to look up the uniqueId IAccessible2
and translate it to a DebugId for an hwnd/childId pair so that
we can find the removed element in the case of destroy.

However, it makes doing this properly for MsaaElement easier.