main.gudl: Work around Skyrim Launcher bug on Wine.
Simulating a click causes an SC_MOVE command. On Windows,
releasing the mouse button immediately ends the move, and then
the button is activated so it's not noticeable. On Wine, the
command gets forwarded to the Linux WM, which can't see our
simulated button release event, so it gets "stuck" in pointer
move mode.

Unfortunately, WM_LBUTTONDOWN causes both of those things, so we
can't just send a message to make the button activate. But, if we
disable the window then SC_MOVE has no effect, while we can still
send a WM_LBUTTONDOWN message to the disabled window.