Add support for UIA providers on non-root MSAA objects.
Apparently, Chrome supports this.

It wasn't previously possible for an ElementIdentifier to have
both an MSAA and UIA provider. In the case of Chrome,
IRawElementProviderFragment is supported, but we don't really
want to use it for navigation. So we assume that whichever
interface gave an id is the "preferred" one.