main.gudl: Add XALIA_SUPPORTED_ONLY environment variable.
This is preparation for enabling Xalia "globally" in Proton,
but not actually enabling it everywhere. This way we can make
general rules in Xalia about when to enable it, such as for all
Unity3D launchers (which should work, but it's not reasonable to
try to maintain a list of all of them).