Win32: Account for HWND reparenting.
Apparently, when a combo box is created, its listbox is created
with the combo box as parent. The listbox is then changed to a
toplevel using SetParent. SetParent may not give any notification
that the desktop window's children have changed.

According to Wine source code, for a visible window, SetParent
will hide and show it, so we should get a notification from that.
Otherwise, we can at least find out once the window becomes visible.

This fixes a combo box in the Expendable launcher.