main.gudl: Treat some ignored winforms controls as supported.
Ideally, "supported" is only for controls we recognize and know
work well, but winforms doesn't really give us type information
beyond the builtin window types it subclasses. If the control
is an msaa "client" (the default), doesn't support UI Automation
(also the default), and can't be tabbed to, we probably don't
need to interact with it at all. Hopefully more "specialized"
controls will actually give us some info to work with and thus
be rejected by this check.